refactor(account): move get account presenter to models and create account api presenter to call all presenters
